More issues with unique_args
The first lines were removed as they were redundant but didn't default args to a hash.

In the event that the line is not called from a mail method at all, we default in the json builder so that the code runs.

The behavior is now as intended.  Args can be set globally or at method level.  At method will override global.